backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
Flask-PyMongo==2.3.0
celery==5.3.6
redis==5.0.1
qiskit==1.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
bcrypt==4.0.1
